src/main/java/ru/job4j/tracker/action/DeleteAction.java
package ru.job4j.tracker.action;

import ru.job4j.tracker.MemTracker;
import ru.job4j.tracker.input.Input;
import ru.job4j.tracker.output.Output;

public class DeleteAction implements UserAction {
    private final Output out;

    public DeleteAction(Output out) {
        this.out = out;
    }

    @Override
    public String name() {
        return "=== Delete Item ===";
    }

    @Override
    public boolean execute(Input input, MemTracker memTracker) {
        int id = input.askInt("Enter items for delete");
        boolean rsl = memTracker.deleted(id);
        if (rsl) {
            out.println("Successful");
        } else {
            out.println("Not successful");
        }
        return true;
    }
}
